tests/requirements.txt
requests==2.31.0
azure-servicebus==7.11.4
pytest==7.4.3